+++ CHANGELOG.md
@@ -1,3 +1,19 @@
+Release 0.1.5 (2020-07-22)
+=====
+* Make the username argument under `emu fork switch` optional. If not specified, it will use the current fork switched to
+* Add `--branch` (`-b`) flag for specifying the branch
+  * This means you must supply `-b` or `--branch` when switching branches, even when supplying the username:
+
+        Old syntax: emu fork switch another_fork branch
+        New syntax: emu fork switch another_fork -b branch
+
+        Old syntax: emu fork switch same_fork new_branch
+        New syntax: emu fork switch -b new_branch
+
+Release 0.1.4 (2020-07-12)
+=====
+* Add `emu device settings` command to open the settings app
+
 Release 0.1.3 (2020-07-06)
 =====
 * Make flags/arguments more robust. Optional non-positional arguments are now supported, as long as they are the last arguments.